requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
pillow==9.5.0
xhtml2pdf==0.2.11
matplotlib
numpy